+++ SistemaEstoque.API/SistemaEstoque.API/Domain/Entity/EntityUsers/Menu.cs
@@ -6,7 +6,7 @@
         public string DescMenu { get; set; }
         public string UrlMenu { get; set; }
         public string Icon { get; set; }
-        public int? MenuFatherId { get; set; }
+        public long? MenuFatherId { get; set; }
         public Guid? EnterpriseId { get; set; }
         public short Active { get; set; }
         public DateTime? CreateDate { get; set; }
